refactor: change decorator name and fix test imports
